npmignore: Don't include the source of tests
We know of no reason that including the source of tests in built npm
packages would be helpful, and we've heard reports that including the
RSA private key fixture as we do in apollo-server-core can trigger
security scans.

Change how we drop tests from "dist" to drop the whole test directory in
case some other files sneak in there.

Add a comment to the one npmignore file that differs from the others.

Fixes #5781.