apollo-server-core: Inline defaultUsageReportingSignature (for graphql@16)
In v3.5 we made AS compatible with the changes in graphql@16, but
apollo-server-core had a dependency on the somewhat inactive
apollo-graphql package which doesn't yet allow graphql@16 in its peer deps.

We should eventually publish a new version of apollo-graphql (eg,
apollo-server-plugin-operation-registry also depends on it) but the
Apollo Server usage reporting plugin is the only consumer of
defaultUsageReportingSignature (and it is not important that the
individual transforms that make it up stay compatible with the building
blocks of other functions like the operation registry signature
function) so inlining seems most natural.

Fixes #5941.