apollo-server-fastify: drop dependency on `fast-json-stringify` (#5988)
This package was being used only to stringify two small constant
objects. It has a major version update available and rather than
evaluate the CHANGELOG, we might as well just not use it.

(Honestly I think it would be fine to just use JSON.stringify here; I
don't see any justification in #1971 for why this particular use case
requires a faster stringification than all the other JSON
stringification we do in Apollo Server. But just in case this matters to
anybody, I went with the fastest possible way of converting a constant
object to string.)

This should replace #5985.