Spell-checking with cspell (#5989)
This commit adds CI using the cspell spell-checker. This is the
command-line form of the VSCode Code Spell Checker extension, which is
now "recommended" if you use this repository. This extension understands
camel-casing in code and has some helpful built-in dictionaries. It is
configured in cspell.yaml; I've set up some nice tweaks like "ignore
things that look like GitHub usernames in CHANGELOG.md".

We have our own dictionary at `cspell-dict.txt` that you can add words
to (either by hand or via the "add word to dictionary" quick fix if
using the extension).  You can also use things like cSpell:ignore,
cSpell:disable-next-line, and cSpell:disable to customize its behavior
on a single file or line if that's helpful.

This commit also fixes a whole bunch of existing spelling errors.  When
easy I've chosen to regard runningwordstogether as a typo to be fixed
rather than as a pattern to allow (the spell checker is fine with
runningWordsTogether).

Like Prettier, we don't run the spell-checker on generated files. To
make configuring this simpler I've moved the output of
graphql-code-generator to a `generated` file just like the protobuf
generated files.