usage reporting: switch to stripIgnoredCharacters
The usage reporting signature algorithm wants to throw away whitespace
and other ignored characters. It currently does this via a hacky set of
regexp replacements.

This PR changes it to use stripIgnoredCharacters which has been in
graphql-js since v14.3.0.

This does have slightly different output. Notably,
stripIgnoredCharacters strips commas, and it includes spaces before
`...` if the previous character is not punctuation. (That's because
otherwise the sequence of tokens `1 ...Foo` would turn into `1...Foo`
which is not valid for the lexer; while this sequence cannot exist in a
legally parsable document, it makes sense that stripIgnoredCharacters
would want to not break the lexer.)

The impact will be that when you upgrade to a version of Apollo Server
with this change, it will appear in Apollo Studio as if all of your
operations have changed even though there is no textual difference.
(Note that Studio already pretty-prints your signature before displaying
it to you which will add back commas where appropriate.)