Remove most web framework integrations
As described in the Roadmap
(https://github.com/apollographql/apollo-server/blob/main/ROADMAP.md),
Apollo Server 4 will replace a static set of hard-coded integrations
with a stable API which allows these integrations to be maintained by
people who actually run them in production.

This commit just removes the old integrations. The new API is not ready
yet, so it's not time to start making the new integration packages, but
this will slim down the size of the version-4 branch while we iterate.

We leave apollo-server-express in place for now, though it will end up
becoming part of the main package.