Remove source of caching API and datasource modules
We plan to replace the cache abstraction with `keyv` (#6045), to move
apollo-datasource and apollo-datasource-rest to their own repos (#6048
and #6049), and replace the top-level dataSources function with a standard
plugin (#6047). So when v4 is done, none of these packages will need to live
in the repo.

This commit just removes their source code; references to the packages now
refer to published versions in the npm registry.

In order to switch these packages from symlink to downloaded, I regenerated
all of package-lock.json.

This tried to upgrade to the latest (transitive) `@types/superagent`
which seems to have broken compiling tests, largely due to
https://github.com/DefinitelyTyped/DefinitelyTyped/pull/57641 That's
probably a good change (it's probably not a good thing that our tests
rely on `dom` being pulled in from a random package) but in order to
avoid worrying about this right now, just pin that package at the latest
version at which we can compile.