docs: Use URL encoding pattern in APQ examples
This alters the format of the `curl` commands we provide for testing/demonstrating APQ usage to URL encode the parameters when sending `GET` requests.

In particular, this commit choses to adopt [`curl`'s `--data-urlencode`] patterns rather than relying on un-URL-encoded passing of values.  While the documentation examples do not use any reserved query-string characters which could be mis-interpreted as a delimiter (i.e., `&`, `=`, etc.) — and thus can get away without it (per [RFC 3986 3.4])  — it likely creates a better precedent to do so.  Much in the same way as it's conceivable that some parameters in a URL don't need [`encodeURIComponent`] 

[`curl`'s `--data-urlencode`]: https://curl.se/docs/manpage.html#--data-urlencode
[RFC 3986 3.4]: https://datatracker.ietf.org/doc/html/rfc3986#section-3.4
[`encodeURIComponent`]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/encodeURIComponent