Remove apollo-server-env for now
apollo-server-env is currently just a thin wrapper around node-fetch
(with our own bespoke typings for it rather than `@types/node-fetch`,
which admittedly is not currently maintained in DefinitelyTyped). We
would like to stop maintaining the wrapper in AS4 (or at least give the
package a more specific name).

Some of these uses are internal and are part of how we represent
*incoming* messages; they will be changed as part of #6038 anyway.
Others are how we talk to external services (eg usage reporting).

For now we're hardcoding all the types and defaults to be `node-fetch`,
which is pretty close to what we had before. Before releasing AS4 we
will figure out the best approach for typing fetcher options.

Part of #6046.