Put plugin-base into types; make CacheScope a pure type
In AS4, `apollo-server-types` was kinda an internal package for sharing
types across the various AS subpackages, whereas
`apollo-server-plugin-base` was an externally-facing package (which
re-exported some types from `apollo-server-types`).

We won't need an "internal-ish" types package in AS4 because the core of
AS will just be one package. We still want to have a separate types
package for plugin authors so they won't have to depend on
`@apollo/server` just for types.

This commit combines the two packages into one; still called
`apollo-server-types` but to be renamed `@apollo/server-types`
eventually.

We will want to do a pass later on making sure this package only
contains truly external-facing types.

While we're at it, we also changed the enum `CacheScope` into a union of
strings type. This means that the types package will truly be a types
package with no runtime exports. This means users will have to write
`'PUBLIC'` instead of `CacheScope.Public` but TypeScript makes this just
as safe anyway. As part of this, remove an `as` in the cache control
plugin and replace it with a real runtime check that validates the same
condition for real.

Part of #6041. Fixes #6110.