Switch from Lerna to npm workspaces
Also switch from nvm to volta. (As part of this, get rid of the direnv
.envrc which said to use npm, though it might be nice to put it back
auto-installing volta later.) We do this because nvm only lets us pin
node and volta lets us pin npm as well, and we'd like to use npm 8.5
which fixes some workspaces-related bugs.

We will later add changesets support for doing the release management
stuff that we used to do with lerna.