Remove path parsing from Apollo Server
Instead of having a magic default path of `/graphql` that is remembered
later inside the server, you just have to mount the middleware returned
from getMiddleware on the path of your choice yourself.

Also remove applyMiddleware which doesn't really offer much of an
improvement over getMiddleware (and which is farther from our eventual
AS4 API).

Part of #6044.