Remove cors, body-parser, and connect support
cors and body-parser are now to be applied by the user alongside the AS
middleware.

We will probably add them back to ApolloServerStandalone later. (Some of
the deleted tests may be useful when that happens.)

Connect is an ancient predecessor of Express that is probably only used
by Meteor these days. We don't need to test against it any more. If
Meteor users need connect support, it should be straightforward to write
connect-specific middleware using the AS4 API, just like for any other
niche web framework.

The middleware can now just return a single function without any runtime
dependency on express!

Part of #6042 and #6043.