Move context creation out of ApolloServer; genericize AS on context
Previously ApolloServer is genericized on the arguments to the context
function rather the actual context object itself. We don't get type
inference into resolvers, inline plugins, etc.

Instead of passing context to the ApolloServer constructor at all, make
it an argument to the integration function. That way the arguments
aren't genericized; they are specific to the function you're calling.
Make the complete context object a parameter to the "execute this
request" function.

We found a lot of places that TContext wasn't threaded through types and
added it.

Now you can do stuff like:

    new ApolloServer<{ foo: number }>({
        Query: {
          n(_parent: any, _args: any, context): number {
            return context.foo;
          },
        },
      },
      plugins: [
        {
          async requestDidStart({ context }) {
            let n: number = context.foo;
          },
        },
        // Plugins declared to be <BaseContext> still work.
        ApolloServerPluginCacheControlDisabled(),
      ],
    });

This also simplifies executeOperation, where you just pass a context
object directly instead of arguments to your (probably already
fake-for-testing) context function.

Fixes #6178.

Paired (in person!!!) with @trevor-scheer.