version-4: Remove docs and renovate file
Don't worry, we still like docs and renovate :)

But we've made a bunch of docs content and infra changes on main (like
getting rid of the package.json) and we'll just pull them in en masse
when we're ready to work on docs here (and this will stop Renovate from
messing with docs here now that we no longer "ignore" it on main).

Also, Renovate only ever reads the config file from main, so let's
delete the one here for now. We already use `matchBaseBranches` on the
main file to have it do different things on version-4 and main.