landing page detection: use a slightly different package
The `accepts` package is just a thin wrapper around `negotiator`, so
switch to the package with the actual code.  Happily, the `@types` type
descriptions for `negotiator` only expect an argument that is "a thing
with a header map" rather than "a Node IncomingMessage" which will be
easier for the framework-agnostic ApolloServer code to construct.

Also, use `negotiator`'s actual "choose which is best" logic instead of
just the "return a sorted list" logic, and add a few more tests.