Make middleware top-level function, no subclassing!
ApolloServerBase is now just ApolloServer. Context function is now a
named argument to the middleware; it's required if your context type is
not the default BaseContext. Note that we differentiate in names between
"contextFunction" (a function returning a context object) and "context"
(the context object itself, eg as an argument to executeOperation).
(We're actually going to change this to context/contextValue in the next
commit.)

Note that when we previously removed direct cors support on this branch,
we removed our discovery that TypeScript sometimes makes you write
`cors<cors.CorsRequest>()` rather than `cors()`. (See
https://github.com/apollographql/apollo-server/issues/5055 where we
figured this out last year.) For some reason the tests were able to
avoid doing that before this change but not any more.