Consistent naming: `context` vs `contextValue`
In AS3 we use `context` for both "a function producing a context value"
and "the context value itself". The previous commit started using
`contextFunction` for the former, but this commit changes to
consistently use `context` for the function and `contextValue` for the
single value. The `contextValue` term matches the name used in
`graphql-js`.

The main place where users write this is providing a context function,
so that name won't change. If they are accessing the context value in a
plugin (on `GraphQLRequestContext` or a named argument to
`willResolveField`) it is now called `contextValue` there.