Factor out common execution code; enforce contravariance
Earlier in version-4 development we tried to ensure that ApolloServer
was contravariant in TContext, so that you can't do this:

     const s: ApolloServer<{}> =
       new ApolloServer<{importantContextField: boolean}>({ ... });
     s.executeOperation({query}, {});

But this broke when we moved all the fields to internals, because
private fields don't have types in .d.ts files and so contravariance
wouldn't be enforced across projects.

This commit uses a more explicit hack to ensure contravariance.
(Hopefully we can soon use an official TS feature for this!)

This was discovered while working on the real work of this commit:
factoring out some shared code between runHttpQuery and
executeOperation. Now there's only one piece of code that sets up a
GraphQLRequestContext!

(Slight behavior change: the response.http for executeOperation now
starts with a content-type: application/json header.)