tests: reorganize integration test suite
Now we only have a single entry point that requires defining a single
setup function for all our integration tests rather than two different
ones for the AS1-era tests and the AS2-era tests.

There is no cleanup function in this interface; if an integration's
tests need to do more cleanup, it can add a serverWillStop plugin to the
ApolloServer.