Merge pull request from GHSA-cpvg-qx3r-pjwv
This new highly recommended feature protects your GraphQL server from
CSRF and XS-Search attacks. It does this by preventing HTTP `GET`
requests that could potentially come from untrusted origins in a web
browser from executing GraphQL operations. Specifically, its default
behavior when enabled is to only execute `GET` requests which contain
`X-Apollo-Operation-Name` or `Apollo-Require-Preflight` headers or which
contain a `Content-Type` header such as `application/json`.

Enabling this feature does not affect any `POST` request (which should
have a `Content-Type: application/json` header), or any `GET` request
from Apollo Client Web, Apollo iOS, or Apollo Kotlin (formerly Apollo
Android). You may need to adjust the configuration of other GraphQL
clients if they send `GET` requests to your server.

This feature will be on by default in Apollo Server 4.

See
https://www.apollographql.com/docs/apollo-server/security/cors#preventing-cross-site-request-forgery-csrf
for more details.

Note that we don't just want to avoid executing the GraphQL operation
--- we also want to avoid running the `context` function (which can read
credentials from the request and could be subject to timing attacks).
The way Apollo Server is structured in its current form means that the
configuration of this feature actually leaks into the top-level
`runHttpQuery` function, because the mechanism for `runHttpQuery` to
learn about the configuration of the `ApolloServer` object is the same
function that runs the `context` function. The refactoring we are doing
as part of Apollo Server 4 fixes this structure so that adding features
like this one will not require making changes to every integration.

This vulnerability was reported by Jeffrey Hofmann; the feature was
designed with advice from Luca Carettoni of Doyensec.