Restore "batteries-include" `ApolloServer` (v4) (#6420)
This commit introduces the "batteries-included" version of ApolloServer v4. This
server manages an express instance internally, meaning it is explicitly not
configurable. This is intended to be a nice getting started experience, but more
complicated use cases should "eject" to the expressMiddleware (i.e. serving
various paths, installing other middlewares, etc.)

The usage is as follows:

```ts
import { ApolloServer, standaloneServer } from "@apollo/server";

interface MyContext {
  token: string;
}

const apolloServerInstance = new ApolloServer<MyContext>({
  typeDefs,
  resolvers,
});
const { url } = await standaloneServer(apolloServerInstance, {
  context: async (req, _res) => ({
    token: await getTokenForRequest(req),
  }),
}).listen({ port: 4000 });

console.log(`ApolloServer listening at: ${url}`);
```

standaloneServer returns an ApolloServerStandalone instance. The intention is to
await the call to listen() on this class. The listen() method accepts the same
arguments that http.Server's listen() method does (in the options form only).

Additionally, this PR introduces a new public function on the ApolloServer
class, addPlugin. This function can be used to add plugins to ApolloServer (only
in its initialized state). This enables us to use the drain http server plugin
with the approach we've taken since the ApolloServer instance (and its plugins)
are instantiated before the httpServer is, which the drain plugin needs a
reference to. This API enables other "standalone server" authors to do the same
if they choose.