tests: remove pluginTestHarness
pluginTestHarness let you test requestPipeline on its own, but in our
new model you can't really use requestPipeline without an ApolloServer.
On the other hand, it's easier to use executeOperation and
executeHTTPGraphQLRequest than it was before. So just use ApolloServers
in these tests and delete pluginTestHarness.

This will make it easier to refactor and ensures that the tests are
testing the right thing, since pluginTestHarness was basically a
reimplementation of the overall logic of requestPipeline.