Embedded Landing Pages  (#6397)
Adds a new "embed" option to the default landing page plugins. This lets you run
Explorer (for production graphs) or Sandbox (for development graphs) directly on
your graph's origin (with the bulk of the code running in an iframe). Among other things,
this lets you interact with your graph without needing to configure CORS.

Co-authored-by: David Glasser <glasser@apollographql.com>