Stop using HttpQueryError in requestPipeline (#6484)
There's a particular check for HttpQueryError that was added in #5287
just to make the "no mutations over GET" plugin work. But given that
that plugin is completely unconfigurable, it might as well just be part
of the request pipeline rather than a plugin. Doing that lets us
simplify sendErrorResponse to no longer have to know about
HttpQueryError for just that use case.

While we're at it, fixed a bug (which exists in AS3) where GET requests
with an invalid operation name threw a ReferenceError in the
checkOperationPlugin instead of the correct error. That's because the
type `GraphQLRequestContextDidResolveOperation` was incorrect:
`operation` is not in fact required at that point. Fix the type (and put
the appropriate `?` in `operation?.operation` in the new code location).

Also remove the unused `isHttpQueryError` function and tweak some
imports.

FOR MIGRATION FILE:
- Fix typing on didResolveOperation argument to accurately note that
  `operation` may be undefined
- Remove unused `isHttpQueryError`