New API to support serverless frameworks (#6505)
The main difference for Apollo Server between serverless and other
frameworks is that serverless frameworks typically need to provide a
request handler *synchronously* on startup. So our model of "you must
call `await server.start()` before your web server listens" doesn't
work.

(The reason we introduced that model (optionally in AS2.22 and required
in AS3) was that previously, it was easy to set up your server in a way
that if it failed to load its schema (eg, from Studio), it would still
start listening and just fail every operation forever. This model lets
your server fail fast if startup fails. However, it doesn't work for
serverless integrations.)

The other minor differences are that the default value of
stopOnTerminationSignals is false for serverless integrations and the
default value of the usage reporting plugin's sendReportsImmediately is
true for serverless integrations; the latter was implemented via the
serverlessFramework option to the plugin serverWillStart API. In both
cases, this merely affected the default values and could be overridden
explicitly to true or false by the user.

The AS3 implementation was based on *subclassing* and overriding a
serverlessFramework() method; if this is true, the ApolloServer
constructor would start the startup process "in the background", logging
errors.

In AS4 we don't subclass ApolloServer at all.  Our new implementation
works by providing a synchronous alternative to `start()` called
`startInBackgroundHandlingStartupErrorsByLoggingAndFailingAllRequests()`.
The length of the name is intended to discourage its use directly by app
writers when they are building non-serverless apps and `await
server.start()` is an option.

Calling this function works similarly to the implicit call at the end of
the AS3 serverless constructor. Additionally, it allows
`server.assertStarted()` to succeed even if startup is still ongoing.

The intended usage goes like this:

For the standalone server, the user does not need to call start at all:

   await startStandaloneServer(server);

(This assumes the API of #6489.)

For a non-serverless middleware, the user should await `start()`. Note
that `expressMiddleware` itself calls `assertStarted`.

  await server.start();
  app.use(expressMiddleware(server));
  app.listen();

For a serverless middleware (not implemented in core), the user doesn't
have to call any start-related function:

  exports.handler = lambdaHandler(server);

`lambdaHandler` should call
`startInBackgroundHandlingStartupErrorsByLoggingAndFailingAllRequests`
for you.

In some cases, a serverless middleware might be implemented in terms of
a non-serverless middleware. For example, one approach to implementing
`lambdaHandler` would be to combine `expressMiddleware` with `express`
and `@vendia/serverless-express`. This is why we have the special case
that `assertStarted` (called by `expressMiddleware`) doesn't throw if
startup is currently taking place in the background.

We continue to have the default values of stopOnTerminationSignals and
sendReportsImmediately be influenced by whether you called start or
startInBackgroundHandlingStartupErrorsByLoggingAndFailingAllRequests. We
do change the serverWillStart option from serverlessFramework to
startedInBackground.

While we're at it, we added a new logged warning if any operations start
executing during the stopping or stopped phase encouraging users to use
a drain plugin.

Note: we add an invocation of the integration tests that combines Express
with this new API. That's not because we think people using Express should
use the API; it just means we were able to ensure that the tests could potentially
work with the new API.  This helped us discover that (in the start-in-background case)
executeHTTPGraphQLRequest throws rather than returns an error response if
startup failed, which we should fix in a subsequent PR.

Fixes #6039.