Inline usage reporting protobuf into main package
Keeping it separate mostly just means we have to worry about version
skew. It doesn't let us avoid the protobufjs dependency (it was just
transitive). Other packages that may need these types (like
`@apollo/gateway`) will have a peer dependency on `@apollo/server`
anyway.

Note that this package consists of .js and .d.ts files that are not
generated from .ts files, so they exist outside of `src`. We ensure they
end up in the published package via a change to `.npmignore`.

As before, the only code that can import the protobuf code should be
tests, `import type`s, or code under `src/plugin` that is only evaluated
if the plugin is actually used.

Fixes #6510.