Get rid of Dispatcher class and associated types (#6529)
In looking over the exported API of Apollo Server, the `AnyFunctionMap`
in the plugin API seemed strange, and its definition already has an old
comment saying that one of its components is not recommended.

This type powers the Dispatcher object and allows it to invoke methods
given a method name in a somewhat typesafe way. But invoking methods via
name not only leads to a lot of complexity in the typing of Dispatcher
--- it also makes it harder to find where the methods are called!

This PR replaces Dispatcher with just keeping lists of listeners around.
In the simplest case of "run them all" we use `await
Promise.all(listeners.map(l => l.hook?.(args)))`. The more complex cases
that involve returning end hooks use new helper functions.

This should mostly be a no-op, though I did make the minor change that
requestDidStart hooks are now called in parallel rather than in series.
Most other hooks were called in series so this will hopefully provide a
minor speed-up in some cases without breaking anything.