Add `cache: "bounded"` configuration option (#6536)
This commit introduces the `cache: "bounded"` option. AS3 has an unbounded cache
by default, which means that a malicious client can take an open-ended amount of
memory in the cache, crashing the server.

Rather than breaking what has been the status quo since the beginning of the
project, we've chosen to add an opt-in option in order to use a bounded cache
with very little configuration. Similar to `csrfPrevention`, we've updated all
examples in our docs to use this `bounded` option.

This option will go away in AS4 when a bounded cache becomes the default.