Remove caching packages (#6541)
Remove apollo-server-caching, apollo-server-cache-redis, and
apollo-server-cache-memcached packages. We will continue to offer bugfixes for
these packages as needed but don't intend to continue active development on
them. Our recommendation going forward will be to use the keyv package along
with our KeyvAdapter from the @apollo/utils.keyvadapter package. Docs updates to
come in a following PR (but coupled with this release).

The KeyValueCache interface and InMemoryLRUCache wrapper around lru-cache (as
well as a few useful cache wrappers like PrefixingKeyValueCache) will continue
to exist in the @apollo/utils.keyvaluecache package. It's worth mentioning that
InMemoryLRUCache is on a newer major version of lru-cache which doesn't permit
the unbounded behavior which we've had by default. In AS 3.9 you can opt into
the bounded cache, and in AS4 it will be the default.